working on DFG

+++ src/main/scala/Chainsaw/DFG/DSPExpressions.scala
@@ -18,7 +18,7 @@ import org.jgrapht.generate._
 import scala.collection.JavaConversions._
 import scala.collection.mutable.ArrayBuffer
 
-case class DSPAssignment[T <: Data](sources: Seq[DSPNode[T]], delays: Seq[Int], var target: DSPNode[T]){
+case class DSPAssignment[T <: Data](sources: Seq[DSPNode[T]], delays: Seq[Double], var target: DSPNode[T]){
   def >=>(that: DSPNode[T]) = {
     target = that
     this
@@ -26,10 +26,10 @@ case class DSPAssignment[T <: Data](sources: Seq[DSPNode[T]], delays: Seq[Int], 
 }
 
 object DSPAssignment {
-  def apply[T <: Data](sources: DSPNode[T], delay: Int, target: DSPNode[T]): DSPAssignment[T] = new DSPAssignment(Seq(sources), Seq(delay), target)
+  def apply[T <: Data](sources: DSPNode[T], delay: Double, target: DSPNode[T]): DSPAssignment[T] = new DSPAssignment(Seq(sources), Seq(delay), target)
 }
 
-case class DSPPath[T <: Data](nodes: ArrayBuffer[DSPNode[T]], delays: ArrayBuffer[Int]){
+case class DSPPath[T <: Data](nodes: ArrayBuffer[DSPNode[T]], delays: ArrayBuffer[Double]){
   def >>(that: DSPNode[T]) = {
     nodes += that
     this
